priority_manager/requirements.txt
click==8.1.7
PyYAML==6.0.2
requests==2.32.3
numpy==2.1.1
msgraph-sdk==1.3.0
msal==1.31.0
